Save all gathered data as nodeinfo1 datatype
This is a proof-of-concept implementation of an experiment that
generates data that meets autoloading requirements.

Specifically, this POC implementation saves all gathered
output from all commands specified in the configuration file as
nodeinfo1 datatype (see api/node_info.go).

For testing and comparing purposes only, the old behavior
is still supported but will be deleted (along with Makefile,
config-old.json, config-new.json) once the changes are fully
reviewed.

The changes in this POC successfully pass static analyzers and
were tested as follows:

    go test ./...
    make
    make check