Set up the filesystem for autoloading
The autoloader agent (jostler) running on M-Lab nodes requires a
datatype schema file and an existing valid path to the directory
where data will be writtent to.  This commit adds a function to
do the above when nodeinfo starts.

Tested the changes with static analyzers and go test ./... -race.

with '#' will be ignored, and an empty message aborts the commit.  #
Dockerfile #	   modified:   main.go #       modified:   main_test.go #